check MCE_SER capability with separate kernel module

+++ setup_mce_kvm.sh
@@ -55,8 +55,22 @@ prepare_mce_kvm() {
 	TARGETGPA=""
 	TARGETHPA=""
 
-	if [ "$ERROR_TYPE" = mce-srao ] ; then
-		check_mce_capability || return 1 # MCE SRAO not supported
+	if ! grep -q mcgcap= $GTMPD/cap_check 2> /dev/null ; then
+		pushd cap_check > /dev/null
+		make check > $GTMPD/cap_check
+		popd > /dev/null
+	fi
+	MCGCAP=$(grep "mcgcap=" $GTMPD/cap_check | cut -f2 -d= | tail -n1)
+	MCE_SER=$(($MCGCAP & (1 << 24)))
+	if [ "$MCE_SER" -gt 0 ] ; then
+		echo "MCE_SER_P supported"
+		MCE_SER_SUPPORTED=true
+	else
+		echo "MCE_SER_P NOT supported"
+		MCE_SER_SUPPORTED=
+		if [ "$ERROR_TYPE" = mce-srao ] ; then
+			return 1
+		fi
 	fi
 
 	prepare_mm_generic || return 1
@@ -76,8 +90,7 @@ prepare_mce_kvm() {
 cleanup_mce_kvm() {
 	save_nr_corrupted_inject
 	all_unpoison
-	echo "pkill -9 -f $_VM_CONSOLE"
-	pkill -9 -f $_VM_CONSOLE
+	[ "$_VM_CONSOLE" ] && pkill -9 -f $_VM_CONSOLE
 	show_guest_console
 	cleanup_mm_generic
 	save_nr_corrupted_unpoison
